bluelines_backend/requirements.txt

faiss-cpu
numpy
PyMuPDF
langchain
langchain-community
langchain-ollama
langchain_openai
lark
groq
python-dotenv
google-cloud-storage

supabase
flask_cors
flask
pinecone
openai